Add the skip test for functional tests (on travis) too...seems those tests weren't actually getting run there at all...
