+++ src/plasTeX/Packages/tests/__init__.py
@@ -18,35 +18,7 @@ py_impl = getattr(platform, 'python_implementation', lambda: None)
 IS_PYPY = py_impl() == 'PyPy'
 
 
-def _real_check_for_binaries():
-    with open('/dev/null', 'wb') as f: # Unix specific (prior to Python 3)
-        subprocess.check_call( ['kpsewhich', '--version'], stdout=f)
-
-_check_for_binaries = _real_check_for_binaries
-
-def _already_checked_for_binaries_and_failed():
-    raise SkipTest("kpsewhich binary not found")
-
-def _already_checked_for_binaries_and_worked():
-    return
-
-def _skip_if_no_binaries():
-    """
-    If the TeX binaries are not available on the PATH in the simple
-    way we use them in these tests, raise unittest's SkipTest
-    exception. This supports testing on Travis CI.
-
-    This is only a partial check and may be slow.
-    """
-    global _check_for_binaries
-    try:
-        _check_for_binaries()
-        _check_for_binaries = _already_checked_for_binaries_and_worked
-    except OSError:
-        _check_for_binaries = _already_checked_for_binaries_and_failed
-        _already_checked_for_binaries_and_failed()
-
-
+from plasTeX.tests import skip_if_no_binaries as _skip_if_no_binaries
 
 if not IS_PYPY:
     import os
